Add cluster operations cheatsheet and update .gitignore for tfplan
This commit introduces a new cheatsheet for cluster operations, providing essential Terraform and Kubernetes commands for managing infrastructure and applications. Additionally, it updates the .gitignore file to include the tfplan file, ensuring that plan files are not tracked in version control. The .gitkeep file in the prod environment has also been removed as it is no longer necessary.